requirements.txt
Flask==3.1.0
SQLAlchemy==2.0.37
pyodbc==5.2.0
gunicorn==20.1.0
urllib3==2.2.1
python-dotenv==1.1.0

